+++ vm/src/tests/mod.rs
@@ -1,5 +1,3 @@
-#[cfg(feature = "cairo-1-hints")]
-use crate::types::layout::CairoLayout;
 use crate::types::layout_name::LayoutName;
 #[cfg(feature = "cairo-1-hints")]
 use crate::vm::errors::cairo_run_errors::CairoRunError;
@@ -108,10 +106,15 @@ fn run_cairo_1_entrypoint(
 
     let mut runner = CairoRunner::new(
         &(contract_class.clone().try_into().unwrap()),
-        CairoLayout::new(LayoutName::all_cairo, None).unwrap(),
-        false,
-        false,
-        false,
+        &Cairo0RunConfig {
+            layout: LayoutName::all_cairo,
+            proof_mode: false,
+            trace_enabled: false,
+            disable_trace_padding: false,
+            ..Default::default()
+        }
+        .run_config()
+        .unwrap(),
     )
     .unwrap();
 
@@ -214,10 +217,15 @@ fn run_cairo_1_entrypoint_with_run_resources(
 ) -> Result<Vec<Felt252>, CairoRunError> {
     let mut runner = CairoRunner::new(
         &(contract_class.clone().try_into().unwrap()),
-        CairoLayout::new(LayoutName::all_cairo, None).unwrap(),
-        false,
-        false,
-        false,
+        &Cairo0RunConfig {
+            layout: LayoutName::all_cairo,
+            proof_mode: false,
+            trace_enabled: false,
+            disable_trace_padding: false,
+            ..Default::default()
+        }
+        .run_config()
+        .unwrap(),
     )
     .unwrap();